Move module code to modules directory
