+++ src/platforms/desktop.c
@@ -14,7 +14,7 @@
 #include "../platform.h"
 #include "../sounds.h"
 
-#include "modules/desktop.h"
+#include "../modules/platforms/desktop.h"
 
 #define FPS 60
 #define FRAME_TIME_LENGTH (1000 / FPS)
